+++ pkg/services/object/search/util.go
@@ -99,9 +99,13 @@ func (c *clientWrapper) searchObjects(ctx context.Context, exec *execCtx, info c
 			return nil, fmt.Errorf("get key for session v2 token: %w", err)
 		}
 	} else if tok := exec.prm.common.SessionToken(); tok != nil {
+		authUser, err := tok.AuthUser()
+		if err != nil {
+			return nil, fmt.Errorf("could not get session auth user: %w", err)
+		}
 		key, err = exec.svc.keyStore.GetKey(&util.SessionInfo{
-			ID:    tok.ID(),
-			Owner: tok.Issuer(),
+			Account: authUser,
+			Owner:   tok.Issuer(),
 		})
 		if err != nil {
 			return nil, err