+++ CHANGELOG.md
@@ -17,10 +17,12 @@ Changelog for NeoFS Node
 ### Changed
 - SN retries notary requests if `insufficient amount of gas` error appears (#3739)
 - Speed up metabase resync by using batch operations (#3804)
+- Session key storage access session keys by account only (#3817)
 
 ### Removed
 
 ### Updated
+- `github.com/nspcc-dev/neofs-sdk-go` module to `v1.0.0-rc.17.0.20260212185528-867eac60c8ca` (#3817)
 
 ### Updating from v0.51.0